Update hitler_was_a_vegetariani_once_met_a_mean_vegan.md
I changed the writing point of view, because it was worded giving advice to people doing outreach. I made it now be more "bot-like", that is, directed to a random stranger. I also removed the "I met a mean vegan once", as I don't think both quite relate. It would be weird for someone to summon the bot about a "mean vegan", and then something about Hitler coming up.